+++ CHANGELOG.md
@@ -1,12 +1,12 @@
 # Changelog
 
-## v0.2.0 - 2024/11/10
+## v0.2.0
 
 * [Rustdesk Client](https://rustdesk.com/docs/en/client/)
 * always download the latest Zoom client version
 * [OBS Studio](https://obsproject.com/)
 
-## v0.1.0 - 2024/11/09
+## v0.1.0
 
 * Debian 12 (Bookworm)
 * GNOME Desktop plus Boxes, Builder and Connections